move checkbox.ts to views folder (#161393)
move checkbox
