+++ src/vs/workbench/browser/parts/views/checkbox.ts
